测试 JS-SDK 支持持续事件回调

+++ lib-webview/src/main/java/com/longrise/android/web/internal/BridgeDelegate.java
@@ -13,8 +13,8 @@ public final class BridgeDelegate implements IBridgeListener {
     private IWebLoadListener mLoadCallback;
 
     private boolean mLoadFailed;
-    private byte mLoadActor;
-    private byte mCurrentLoadStatus = -1;
+    private int mLoadActor;
+    private int mCurrentLoadStatus = -1;
 
     @Override
     public boolean beforeUrlLoading(String url) {
